+++ src/backend/controllers.js
@@ -20,7 +20,7 @@ async function getTasks(req, res) {
             return res.status(200).json(result.filter(task => task.title.toLowerCase().includes(q.toLowerCase())));
         }
         
-            return res.status(200).json(result);
+            return res.status(200).json(result.sort((nextTask, prevTask) => prevTask.creationDate - nextTask.creationDate));
     });
 }
