+++ src/frontend/app/components/base/custom_button/index.js
@@ -1,0 +1,1 @@
+export { CustomButton } from './components';
