+++ src/frontend/app/components/base/modal/index.js
@@ -1,1 +1,2 @@
 export { Modal } from "./components";
+export { openModal, closeModal } from "./helpers";
