--- src/frontend/app/components/base/sort_elem/index.js
@@ -1,1 +1,0 @@
-export {SortElem} from './components';
