+++ src/frontend/app/components/footer/aside/components.js
@@ -1,12 +1,14 @@
 import styles from './styles.module.scss';
-import {SortElem} from "../../base/sort_elem";
+import {
+    CustomButton
+} from "../../base/custom_button";
 
 export function Aside() {
     const aside = document.createElement('aside');
 
     aside.classList.add(styles.aside);
 
-    aside.append(SortElem("Help", styles.help));
+    aside.append(CustomButton("Help", styles.help));
 
     return aside;
 }
