+++ src/frontend/app/components/main/section/search_panel/add_task_form/styles.module.scss
@@ -10,4 +10,7 @@
     .description::placeholder {
         color: #ffffff;
     }
+    .add-task{
+        width: 130px;
+    }
 }
