+++ src/frontend/app/components/main/section/search_panel/keyword_search_bar/index.js
