+++ src/frontend/app/components/main/side_panel/styles.module.scss
@@ -17,7 +17,7 @@
     }
 }
 
-.sort-elem {
+.custom-button {
     height: 15%;
     margin-top: 10px;
 }
